criando serviço de tipo

+++ book-room/quarto/src/main/java/br/com/book/room/quarto/domain/core/AmenidadesLocalidadeId.java
@@ -2,5 +2,5 @@ package br.com.book.room.quarto.domain.core;
 
 import java.io.Serializable;
 
-public record AmenidadesLocalidadeId(Integer idLocalidade, Integer idAmenidade) implements Serializable {
+public record AmenidadesLocalidadeId(Long idLocalidade, Long idAmenidade) implements Serializable {
 }
